% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachments.R
\name{attachments}
\alias{attachments}
\title{Get attachments informations}
\usage{
attachments(data = NULL)
}
\arguments{
\item{data}{tibble of datasets as given by \code{datasets()} or \code{NULL} (default)}
}
\value{
a tibble of datafiles informations
}
\description{
return information on all datafiles or on
can use a \code{datasets()} command as entry
}
\examples{
\donttest{
dfs <- datafiles()
}
}
