% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{build_attributes}
\alias{build_attributes}
\title{build_attributes}
\usage{
build_attributes(columns)
}
\arguments{
\item{columns;}{as returned by \code{columns()}}
}
\description{
return attributes to attach to downloaded dataframe
}
\keyword{internal}
