% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{build_url}
\alias{build_url}
\title{Build url}
\usage{
build_url(path, default_query = list(), query = list())
}
\arguments{
\item{path}{the api route}

\item{default_query}{the default parameters for the query as a vector}

\item{query}{the users parameters for the query as a vector}
}
\value{
the complete url to call
}
\description{
build url from all parameters
}
\examples{
build_url("/datasets", query = c(page = 1, pageSize = 1))
build_url("/datasets",
  default_query = c(default = "default"),
  query = c(page = 1, pageSize = 1)
)
}
\keyword{internal}
