% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millesimes.R
\name{columns}
\alias{columns}
\title{Get columns count on millesimes}
\usage{
columns(data)
}
\arguments{
\item{data}{a datafiles tibble (as returned by the \code{datafiles()} function)}
}
\value{
a tibble with columns found in input datafiles and occurrence for
each column
}
\description{
This command return the list of columns founded in the input tibble with
the count. It allows you to verify millesimes are coherent
}
\examples{
datafiles() \%>\% columns()
}
