% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafiles.R
\name{datafiles}
\alias{datafiles}
\title{Get datafiles informations}
\usage{
datafiles(data = NULL)
}
\arguments{
\item{data}{tibble of datasets as given by \code{datasets()} or \code{NULL} (default)}
}
\value{
a tibble of datafiles informations
}
\description{
return information on all datafiles or on
can use a \code{datasets()} command as entry
}
\examples{
dfs <- datafiles()
}
