% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets}
\alias{datasets}
\title{Get datasets informations}
\usage{
datasets(data = NULL)
}
\arguments{
\item{data}{a datafiles tibble (as returned by the \code{datafiles()} function) or NULL (default)}
}
\value{
a tibble of datasets information
}
\description{
Load datasets information in a tibble if needed and return it
}
\examples{
datasets()
}
