% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dido-search.R
\name{dido_search}
\alias{dido_search}
\title{Return row matching pattern}
\usage{
dido_search(data, pattern)
}
\arguments{
\item{data}{Tibble of datasets or datafiles as returned by datasets()
and datafiles()}

\item{pattern}{string used to filter the dataset and idbank list}
}
\value{
the datafiles or datasets table filtered with the pattern
}
\description{
Return row matching pattern
}
\examples{
dido_search(datasets(), "drom")
}
