% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didor.R
\docType{package}
\name{didor-package}
\alias{didor-package}
\alias{_PACKAGE}
\title{didor: Tools to Explore and Download CGDD/SDES Opendata}
\description{
Tools to download data from the CGDD/SDES database \url{https://data.statistiques.developpement-durable.gouv.fr/dido/api/v1/apidoc.html} together with search and manipulation utilities.
}
\author{
\strong{Maintainer}: Nicolas Chuche \email{nicolas.chuche@developpement-durable.gouv.fr}

}
\keyword{internal}
