% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-path.R
\name{get_api_base_path}
\alias{get_api_base_path}
\title{Get API base path}
\usage{
get_api_base_path()
}
\value{
the api base path
}
\description{
Get the API base path. Can be configured with \code{set_api_base_path} and with
\code{DIDOLITE_API_BASE_PATH} env var. Order is :
\enumerate{
\item if set_api_base_path has been used
\item \code{DIDOLITE_API_BASE_PATH}
\item the default default_api_base_path
}
}
\examples{
get_api_base_path()
}
