% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-attachments.R
\name{get_attachments}
\alias{get_attachments}
\title{download attachments}
\usage{
get_attachments(data, dest = NULL)
}
\arguments{
\item{data}{tibble of datasets as given by \code{attachments()}.
Must contains an \code{id} column}

\item{dest}{directory to write files}
}
\description{
download attachments
}
\examples{
datasets() \%>\%
  dido_search("drom") \%>\%
  attachments() \%>\%
  get_attachments(dest = tempdir())
}
