% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{get_attributes}
\alias{get_attributes}
\title{get_attributes}
\usage{
get_attributes(data)
}
\arguments{
\item{data}{a data frame as returned by get_data}
}
\value{
a named list of named list of attributes by column names
}
\description{
used internally to reapply attributes on data
}
\keyword{internal}
