% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dido_api.R
\name{get_metadata}
\alias{get_metadata}
\title{Get metadata}
\usage{
get_metadata(url = NULL)
}
\arguments{
\item{url}{empty, url or filename. If no argument is passed, the default
url is used. file must be compatible with the dido api format for
https://data.statistiques.developpement-durable.gouv.fr/dido/api/v1/apidoc.html#/datasets/paginate_datasets
datasets as described on}
}
\value{
nothing
}
\description{
Download information from api of from file passed as argument
Load those data as tibble
}
\examples{
\dontrun{
get_metadata("datasets.json")
}
}
