% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{http_get}
\alias{http_get}
\title{Run HTTP GET request}
\usage{
http_get(url, as = NULL)
}
\arguments{
\item{url}{the url}

\item{as}{text/NULL, return as text or as is}
}
\value{
a didor_http_response with two fields:
\itemize{
\item parsed content with the \code{as} parameter
\item headers
}
}
\description{
Run HTTP GET request
}
\examples{
http_get(
  "https://data.statistiques.developpement-durable.gouv.fr/dido/api/v1/datasets?page=1&pageSize=10",
  as = "text"
)
}
\keyword{internal}
