% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millesimes.R
\name{last_millesime}
\alias{last_millesime}
\title{Get last millesime}
\usage{
last_millesime(data)
}
\arguments{
\item{data}{a datafiles tibble (as returned by the \code{datafiles()} function)}
}
\value{
a tibble with the last millesimes for all input datafiles
}
\description{
Get last millesime
}
\examples{
datafiles() \%>\% last_millesime()
}
