% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millesimes.R
\name{millesimes}
\alias{millesimes}
\title{Get millesimes informations}
\usage{
millesimes(data)
}
\arguments{
\item{data}{a tibble issued from \code{datafiles()}.}
}
\value{
a tibble of millesimes informations
}
\description{
Get millesimes informations
}
\examples{
datafiles() \%>\% dido_search("drom") \%>\% millesimes()
}
