% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_path.R
\name{set_api_base_path}
\alias{set_api_base_path}
\title{Set API base path}
\usage{
set_api_base_path(path = NULL)
}
\arguments{
\item{path}{the path to set or NULL (default value).
if set to non \code{NULL}, set the path
if \code{NULL}}
}
\value{
nothing
}
\description{
set a new API base path to use
}
\examples{
# to set some new path
set_api_base_path("http://localhost/api/")
}
