% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{set_attributes}
\alias{set_attributes}
\title{set_attributes}
\usage{
set_attributes(data, metadata)
}
\arguments{
\item{data}{a dataframe}

\item{metadata}{named list as returned by \code{build_attributes()}}
}
\value{
a dataframe with attributes on columns
}
\description{
add information to downloaded dataframe as attributes
\itemize{
\item description
\item unit
\item type : concatenation of type and format as returned by API
}
}
\keyword{internal}
