% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-path.R
\name{set_api_base_path}
\alias{set_api_base_path}
\title{Set API base path}
\usage{
set_api_base_path(path = NULL)
}
\arguments{
\item{path}{or NULL the path to set or NULL to use to default path.
\itemize{
\item if set to non \code{NULL}, set the value as path
\item if \code{NULL} set the default path
\verb{https://data.statistiques.developpement-durable.gouv.fr/dido/api/v1}
}

@details this function delete the cache.}
}
\value{
the previous path
}
\description{
set a new API base path to use
}
\examples{
# to set some new path
set_api_base_path("http://localhost/api/")
}
